backend/requirements.txt
# Core Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.0

# Redis
redis==5.2.0
redis-om==0.3.2

# Validation & Serialization
pydantic==2.9.2
pydantic-settings==2.6.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# WebSocket
websockets==13.1

# Utilities
python-dotenv==1.0.1
httpx==0.27.2

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
black==24.10.0
ruff==0.8.4
mypy==1.13.0

